Fix graceful shutdown of the service
By default systemd sends `SIGTERM` signal when stopping the service.
Previously the service logic was adding CTRLC handler but it was not
called because `SIGINT` is not the signal systemd was using for shutting
down the service. This adjusts systemd service definition file to use
`SIGINT` signal instead. Now, when running `systemd stop` on the
service, we can verify that the CTRLC handler is being executed.
This adjustment would fix the issue with the service not releasing
NZXT USB device handle between service restarts. That is no longer the
issue though because I switched to lm_sensors based approach. Still, this
change makes the SIGINT handler work as expected.